modifying Gui interfaces
	  setting size_t instead of int for height and width

+++ interfaces/IGui.hpp
@@ -10,7 +10,7 @@ public:
   virtual ~IGui() { };
 
   /* createWindows -1 for failure */
-  virtual int createWindows(std::pair<int, int>) = 0;
+  virtual int createWindows(std::pair<size_t, size_t>) = 0;
 
   /* launchMenu return 0 for success and -1 for failure */
   virtual int launchMenu() = 0;
